requirements.txt
Django>=5.1
djangorestframework>=3.16
django-cors-headers>=4.9
whitenoise>=6.12
